Struct bonsaidb_core::admin::User

pub struct User {
    pub username: String,
    pub groups: Vec<u64>,
    pub roles: Vec<u64>,
    pub argon_hash: Option<SensitiveString>,
}

A user that can authenticate with BonsaiDb.

Fields
username: String

The name of the role. Must be unique.

groups: Vec<u64>

The IDs of the user groups this user belongs to.

roles: Vec<u64>

The IDs of the roles this user has been assigned.

argon_hash: Option<SensitiveString>

The user’s stored password hash.

This field is not feature gated to prevent losing stored passwords if the password-hashing feature is disabled and then re-enabled and user records are updated in the meantime.

Implementations

impl User

pub fn assume_identity<'name, Storage: StorageConnection>(
name_or_id: impl Nameable<'name, u64>,
storage: &Storage
) -> Result<Storage::Authenticated, Error>

pub async fn assume_identity_async<'name, Storage: AsyncStorageConnection>(
name_or_id: impl Nameable<'name, u64> + Send,
storage: &Storage
) -> Result<Storage::Authenticated, Error>

pub fn default_with_username(username: impl Into<String>) -> Self

Returns a default user with the given username.

pub fn effective_permissions<C: Connection>(
&self,
admin: &C,
inherit_permissions: &Permissions
) -> Result<Permissions, Error>

Calculates the effective permissions based on the groups and roles this user is assigned.