Module bonsaidb_core::api

Traits for tailoring a server.

Structs

ApiName

The qualified name of an Api.

Enums

Infallible

An Error type that can be used in within an Api definition.

Traits

Api

An API request type. This trait is used by BonsaiDb’s server to allow a client to send a request of this type, and the server can respond with a Result<Api::Response,Api::Error>.

ApiError

An error that can be used within a Api definition.

Type Definitions

ApiResult

The result of executing a custom API call.

Derive Macros

Api

Derives the bonsaidb::core::api::Api trait. #[api(name = "Name", authority = "Authority", response = ResponseType, error = ErrorType, core = bonsaidb::core)] authority, response, error and core are optional