Struct bonsaidb_core::connection::AsyncCollection

pub struct AsyncCollection<'a, Cn, Cl> { /* private fields */ }

Interacts with a collection over a Connection.

These examples in this type use this basic collection definition:

use bonsaidb_core::{schema::Collection, Error};
use serde::{Deserialize, Serialize};

#[derive(Debug, Serialize, Deserialize, Default, Collection)]
#[collection(name = "MyCollection")]
pub struct MyCollection {
    pub rank: u32,
    pub score: f32,
}

Implementations

impl<'a, Cn, Cl> AsyncCollection<'a, Cn, Cl> where
Cn: AsyncConnection,
Cl: Collection,

pub async fn push(
&self,
item: &<Cl as SerializedCollection>::Contents
) -> Result<CollectionHeader<Cl::PrimaryKey>, Error> where
Cl: SerializedCollection,

Adds a new Document<Cl> with the contents item.

Automatic ID Assignment

This function calls SerializedCollection::natural_id() to try to retrieve a primary key value from item. If an id is returned, the item is inserted with that id. If an id is not returned, an id will be automatically assigned, if possible, by the storage backend, which uses the Key trait to assign ids.

let inserted_header = db
    .collection::<MyCollection>()
    .push(&MyCollection::default())
    .await?;
println!(
    "Inserted id {} with revision {}",
    inserted_header.id, inserted_header.revision
);

pub async fn push_bytes<B: Into<Bytes> + Send>(
&self,
contents: B
) -> Result<CollectionHeader<Cl::PrimaryKey>, Error>

Adds a new Document<Cl> with the contents.

Automatic ID Assignment

An id will be automatically assigned, if possible, by the storage backend, which uses the Key trait to assign ids.

let inserted_header = db.collection::<MyCollection>().push_bytes(vec![]).await?;
println!(
    "Inserted id {} with revision {}",
    inserted_header.id, inserted_header.revision
);

pub async fn insert<PrimaryKey>(
&self,
id: &PrimaryKey,
item: &<Cl as SerializedCollection>::Contents
) -> Result<CollectionHeader<Cl::PrimaryKey>, Error> where
Cl: SerializedCollection,
PrimaryKey: for<'k> KeyEncoding<'k, Cl::PrimaryKey> + ?Sized,

Adds a new Document<Cl> with the given id and contents item.

let inserted_header = db
    .collection::<MyCollection>()
    .insert(&42, &MyCollection::default())
    .await?;
println!(
    "Inserted id {} with revision {}",
    inserted_header.id, inserted_header.revision
);

pub async fn insert_bytes<PrimaryKey, B: Into<Bytes> + Send>(
&self,
id: &PrimaryKey,
contents: B
) -> Result<CollectionHeader<Cl::PrimaryKey>, Error> where
PrimaryKey: for<'k> KeyEncoding<'k, Cl::PrimaryKey> + ?Sized,

Adds a new Document<Cl> with the the given id and contents.

let inserted_header = db
    .collection::<MyCollection>()
    .insert_bytes(&42, vec![])
    .await?;
println!(
    "Inserted id {} with revision {}",
    inserted_header.id, inserted_header.revision
);

pub async fn update<D: Document<Cl> + Send + Sync>(
&self,
doc: &mut D
) -> Result<(), Error>

Updates an existing document. Upon success, doc.revision will be updated with the new revision.

if let Some(mut document) = db.collection::<MyCollection>().get(&42).await? {
    // modify the document
    db.collection::<MyCollection>().update(&mut document);
    println!("Updated revision: {:?}", document.header.revision);
}

pub async fn overwrite<D: Document<Cl> + Send + Sync>(
&self,
doc: &mut D
) -> Result<(), Error>

Overwrites an existing document, or inserts a new document. Upon success, doc.revision will be updated with the new revision information.

if let Some(mut document) = db.collection::<MyCollection>().get(&42).await? {
    // modify the document
    db.collection::<MyCollection>().overwrite(&mut document);
    println!("Updated revision: {:?}", document.header.revision);
}

pub async fn get<PrimaryKey>(
&self,
id: &PrimaryKey
) -> Result<Option<OwnedDocument>, Error> where
PrimaryKey: for<'k> KeyEncoding<'k, Cl::PrimaryKey> + ?Sized,

Retrieves a Document<Cl> with id from the connection.

if let Some(doc) = db.collection::<MyCollection>().get(&42).await? {
    println!(
        "Retrieved bytes {:?} with revision {}",
        doc.contents, doc.header.revision
    );
    let deserialized = MyCollection::document_contents(&doc)?;
    println!("Deserialized contents: {:?}", deserialized);
}

pub async fn get_multiple<'id, DocumentIds, PrimaryKey, I>(
&self,
ids: DocumentIds
) -> Result<Vec<OwnedDocument>, Error> where
DocumentIds: IntoIterator<Item = &'id PrimaryKey, IntoIter = I> + Send + Sync,
I: Iterator<Item = &'id PrimaryKey> + Send + Sync,
PrimaryKey: for<'k> KeyEncoding<'k, Cl::PrimaryKey> + 'id + ?Sized,

Retrieves all documents matching ids. Documents that are not found are not returned, but no error will be generated.

for doc in db
    .collection::<MyCollection>()
    .get_multiple(&[42, 43])
    .await?
{
    println!("Retrieved #{} with bytes {:?}", doc.header.id, doc.contents);
    let deserialized = MyCollection::document_contents(&doc)?;
    println!("Deserialized contents: {:?}", deserialized);
}

pub fn list<PrimaryKey, R>(
&'a self,
ids: R
) -> AsyncList<'a, Cn, Cl, PrimaryKey> where
R: Into<RangeRef<'a, Cl::PrimaryKey, PrimaryKey>>,
PrimaryKey: for<'k> KeyEncoding<'k, Cl::PrimaryKey> + PartialEq + ?Sized,
Cl::PrimaryKey: Borrow<PrimaryKey> + PartialEq<PrimaryKey>,

Retrieves all documents matching the range of ids.

for doc in db
    .collection::<MyCollection>()
    .list(42..)
    .descending()
    .limit(20)
    .await?
{
    println!("Retrieved #{} with bytes {:?}", doc.header.id, doc.contents);
    let deserialized = MyCollection::document_contents(&doc)?;
    println!("Deserialized contents: {:?}", deserialized);
}

pub fn list_with_prefix<PrimaryKey>(
&'a self,
prefix: &'a PrimaryKey
) -> AsyncList<'a, Cn, Cl, PrimaryKey> where
PrimaryKey: IntoPrefixRange<'a, Cl::PrimaryKey> + for<'k> KeyEncoding<'k, Cl::PrimaryKey> + PartialEq + ?Sized,
Cl::PrimaryKey: Borrow<PrimaryKey> + PartialEq<PrimaryKey>,

Retrieves all documents with ids that start with prefix.

use bonsaidb_core::{
    connection::AsyncConnection,
    document::OwnedDocument,
    schema::{Collection, Schematic, SerializedCollection},
    Error,
};
use serde::{Deserialize, Serialize};

#[derive(Debug, Serialize, Deserialize, Default, Collection)]
#[collection(name = "MyCollection", primary_key = String)]
pub struct MyCollection;

async fn starts_with_a<C: AsyncConnection>(db: &C) -> Result<Vec<OwnedDocument>, Error> {
    db.collection::<MyCollection>().list_with_prefix("a").await
}

pub fn all(&'a self) -> AsyncList<'a, Cn, Cl, Cl::PrimaryKey>

Retrieves all documents.

for doc in db.collection::<MyCollection>().all().await? {
    println!("Retrieved #{} with bytes {:?}", doc.header.id, doc.contents);
    let deserialized = MyCollection::document_contents(&doc)?;
    println!("Deserialized contents: {:?}", deserialized);
}

pub async fn delete<H: HasHeader + Send + Sync>(
&self,
doc: &H
) -> Result<(), Error>

Removes a Document from the database.

if let Some(doc) = db.collection::<MyCollection>().get(&42).await? {
    db.collection::<MyCollection>().delete(&doc).await?;
}