Type Definition bonsaidb_core::connection::GroupedReductions

pub type GroupedReductions<V> = Vec<MappedValue<<V as View>::Key, <V as View>::Value>>;

This type is the result of reduce_grouped(). It is a list of all matching keys and the reduced value of all mapped entries for that key.