Trait bonsaidb_core::key::IntoPrefixRange

pub trait IntoPrefixRange<'a, TOwned>: PartialEqwhere
    TOwned: Borrow<Self> + PartialEq<Self>,{
    fn to_prefix_range(&'a self) -> RangeRef<'a, TOwned, Self>;
}

A type that can be used as a prefix range in range-based queries.

Required Methods

fn to_prefix_range(&'a self) -> RangeRef<'a, TOwned, Self>

Returns the value as a prefix-range, which will match all values that start with self.