Module bonsaidb_core::networking

Types for implementing the BonsaiDb network protocol.

Structs

AlterUserPermissionGroupMembership

Alter’s a user’s membership in a permission group.

AlterUserRoleMembership

Alter’s a user’s role

ApplyTransaction

Applies a transaction.

AssumeIdentity

Assume an identity.

Authenticate

Authenticate as a user.

Compact

Compacts the entire database.

CompactCollection

Compacts the collection.

CompactKeyValueStore

Compacts the key-value store.

Count

Counts the number of documents in the specified range.

CreateDatabase

Creates a database.

CreateSubscriber

Creates a PubSub Subscriber

CreateUser

Creates a user.

DeleteDatabase

Deletes the database named name

DeleteDocs

Deletes the associated documents resulting from the view query.

DeleteUser

Deletes a user.

ExecuteKeyOperation

Excutes a key-value store operation.

Get

Retrieve a single document.

GetMultiple

Retrieve multiple documents.

LastTransactionId

Queries the last transaction id.

List

Retrieve multiple documents.

ListAvailableSchemas

Lists available schemas.

ListDatabases

Lists all databases.

ListExecutedTransactions

Lists executed transactions.

ListHeaders

Retrieve multiple document headers.

MessageReceived

A PubSub message was received.

Payload

A payload with an associated id.

Publish

Publishes payload to all subscribers of topic.

PublishToAll

Publishes payload to all subscribers of all topics.

Query

Queries a view.

QueryWithDocs

Queries a view with the associated documents.

Reduce

Reduces a view.

ReduceGrouped

Reduces a view, grouping the reduced values by key.

SetUserPassword

Set’s a user’s password.

SubscribeTo

Subscribes subscriber_id to messages for topic.

UnregisterSubscriber

Unregisters the subscriber.

UnsubscribeFrom

Unsubscribes subscriber_id from messages for topic.

Enums

Error

A networking error.

Constants

CURRENT_PROTOCOL_VERSION

The current protocol version.