Module bonsaidb_core::schema

Types for defining database schema.

Re-exports

pub use self::view::map::Key;

pub use self::view::map::Map;

pub use self::view::map::MappedDocument;

pub use self::view::map::MappedValue;

pub use self::view::MapResult;

pub use self::view::View;

Modules

view
Types for defining map/reduce-powered Views.

Structs

Authority
The owner of a schema item. This should represent the company, group, or individual that created the item in question. This value is used for namespacing. Changing this after values are in use is not supported without manual migrations at this time.

CollectionDocument
A document with serializable contents.

CollectionName
The name of a Collection.

InvalidNameError
An invalid name was used in a schema definition.

Name
A valid schema name. Must be alphanumeric (a-zA-Z9-0) or a hyphen (-). Cloning this structure shares the underlying string data, regardless of whether it’s a static string literal or an owned String.

SchemaName
The name of a Schema.

Schematic
A collection of defined collections and views.

ViewName
The name of a View.

Enums

NamedReference
A reference to a collection that has a unique name view.

Traits

Collection
A namespaced collection of Document<Self> items and views.

NamedCollection
A collection with a unique name column.

Schema
Defines a group of collections that are stored into a single database.