Struct bonsaidb_core::schema::Schematic

pub struct Schematic {
    pub name: SchemaName,
    /* private fields */
}

A collection of defined collections and views.

Fields
name: SchemaName

The name of the schema this was built from.

Implementations

impl Schematic

pub fn from_schema<S: Schema + ?Sized>() -> Result<Self, Error>

Returns an initialized version from S.

pub fn define_collection<C: Collection + 'static>(
&mut self
) -> Result<(), Error>

Adds the collection C and its views.

pub fn define_view<V: ViewSchema<View = V> + SerializedView + Clone + 'static>(
&mut self,
view: V
) -> Result<(), Error>

Adds the view V.

pub fn define_view_with_schema<V: SerializedView + 'static, S: ViewSchema<View = V> + 'static>(
&mut self,
view: V,
schema: S
) -> Result<(), Error>

Adds the view V.

pub fn contains_collection<C: Collection + 'static>(&self) -> bool

Returns true if this schema contains the collection C.

pub fn contains_collection_name(&self, collection: &CollectionName) -> bool

Returns true if this schema contains the collection C.

pub fn next_id_for_collection(
&self,
collection: &CollectionName,
id: Option<DocumentId>
) -> Result<DocumentId, Error>

Returns the next id in sequence for the collection, if the primary key type supports the operation and the next id would not overflow.

pub fn view_by_name(&self, name: &ViewName) -> Result<&dyn Serialized, Error>

Looks up a view::Serialized by name.

pub fn view<V: View + 'static>(&self) -> Result<&dyn Serialized, Error>

Looks up a view::Serialized through the the type V.

pub fn views(&self) -> impl Iterator<Item = &dyn Serialized>

Iterates over all registered views.

pub fn views_in_collection(
&self,
collection: &CollectionName
) -> Option<Vec<&dyn Serialized>>

Iterates over all views that belong to collection.

pub fn eager_views_in_collection(
&self,
collection: &CollectionName
) -> Option<Vec<&dyn Serialized>>

Iterates over all views that are eagerly updated that belong to collection.

pub fn encryption_key_for_collection(
&self,
collection: &CollectionName
) -> Option<&KeyId>

Returns a collection’s default encryption key, if one was defined.

pub fn collections(&self) -> Vec<CollectionName>

Returns a list of all collections contained in this schematic.