Trait bonsaidb_core::schema::Collection

pub trait Collection: Debug + Send + Sync {
    fn collection_name() -> Result<CollectionName, InvalidNameError>;

    fn define_views(schema: &mut Schematic) -> Result<(), Error>;

    fn encryption_key() -> Option<KeyId> { ... }

    fn serializer() -> CollectionSerializer { ... }

    fn get<'life0, 'async_trait, C: Connection>(
        id: u64, 
        connection: &'life0 C
    ) -> Pin<Box<dyn Future<Output = Result<Option<CollectionDocument<Self>>, Error>> + Send + 'async_trait>>
    where
        Self: Serialize + for<'de> Deserialize<'de>,
        C: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn insert_into<'life0, 'async_trait, Cn: Connection>(
        self, 
        connection: &'life0 Cn
    ) -> Pin<Box<dyn Future<Output = Result<CollectionDocument<Self>, InsertError<Self>>> + Send + 'async_trait>>
    where
        Self: Serialize + for<'de> Deserialize<'de> + 'static,
        Cn: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
}

A namespaced collection of Document<Self> items and views.

Required methods

fn collection_name() -> Result<CollectionName, InvalidNameError>

The Id of this collection.

fn define_views(schema: &mut Schematic) -> Result<(), Error>

Defines all Views in this collection in schema.

Provided methods

fn encryption_key() -> Option<KeyId>

If a KeyId is returned, this collection will be stored encrypted at-rest using the key specified.

fn serializer() -> CollectionSerializer

Returns the serializer to use when accessing and storing the document’s contents. If you only interact with a document’s bytes directly, this has no effect.

fn get<'life0, 'async_trait, C: Connection>(
id: u64,
connection: &'life0 C
) -> Pin<Box<dyn Future<Output = Result<Option<CollectionDocument<Self>>, Error>> + Send + 'async_trait>> where
Self: Serialize + for<'de> Deserialize<'de>,
C: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

Gets a CollectionDocument with id from connection.

fn insert_into<'life0, 'async_trait, Cn: Connection>(
self,
connection: &'life0 Cn
) -> Pin<Box<dyn Future<Output = Result<CollectionDocument<Self>, InsertError<Self>>> + Send + 'async_trait>> where
Self: Serialize + for<'de> Deserialize<'de> + 'static,
Cn: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

Inserts this value into the collection, returning the created document.