Trait bonsaidb_core::schema::NamedCollection

pub trait NamedCollection: Collection + Unpin {
    type ByNameView: SerializedView<Key = String>;
    fn load<'name, N: Nameable<'name, Self::PrimaryKey> + Send + Sync, C: Connection>(
        id: N, 
        connection: &C
    ) -> Result<Option<CollectionDocument<Self>>, Error>
    where
        Self: SerializedCollection + Sized + 'static,
    { ... }

    fn load_async<'name, 'life0, 'async_trait, N, C>(
        id: N, 
        connection: &'life0 C
    ) -> Pin<Box<dyn Future<Output = Result<Option<CollectionDocument<Self>>, Error>> + Send + 'async_trait>>
    where
        Self: SerializedCollection + Sized + 'static,
        'name: 'async_trait,
        N: 'async_trait + Nameable<'name, Self::PrimaryKey> + Send + Sync,
        C: 'async_trait + AsyncConnection,
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }

    fn entry<'connection, 'name, N: Into<NamedReference<'name, Self::PrimaryKey>> + Send + Sync, C: Connection>(
        id: N, 
        connection: &'connection C
    ) -> Entry<'connection, 'name, C, Self, (), ()>
    where
        Self: SerializedCollection + Sized,
    { ... }

    fn entry_async<'connection, 'name, N: Into<NamedReference<'name, Self::PrimaryKey>> + Send + Sync, C: AsyncConnection>(
        id: N, 
        connection: &'connection C
    ) -> AsyncEntry<'connection, 'name, C, Self, (), ()>
    where
        Self: SerializedCollection + Sized,
    { ... }

    fn load_document<'name, N: Nameable<'name, Self::PrimaryKey> + Send + Sync, C: Connection>(
        name: N, 
        connection: &C
    ) -> Result<Option<OwnedDocument>, Error>
    where
        Self: SerializedCollection + Sized,
    { ... }

    fn load_document_async<'name, 'life0, 'async_trait, N, C>(
        name: N, 
        connection: &'life0 C
    ) -> Pin<Box<dyn Future<Output = Result<Option<OwnedDocument>, Error>> + Send + 'async_trait>>
    where
        Self: SerializedCollection + Sized,
        'name: 'async_trait,
        N: 'async_trait + Nameable<'name, Self::PrimaryKey> + Send + Sync,
        C: 'async_trait + AsyncConnection,
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
}

A collection with a unique name column.

Finding a document by unique name

if let Some(doc) = MyCollection::load("unique name", &db)? {
    println!(
        "Retrieved revision {} with deserialized contents: {:?}",
        doc.header.revision, doc.contents
    );
}

Load accepts either a string or a DocumentId. This enables building methods that accept either the unique ID or the unique name:

if let Some(doc) = MyCollection::load(42, &db)? {
    println!(
        "Retrieved revision {} with deserialized contents: {:?}",
        doc.header.revision, doc.contents
    );
}

Executing an insert or update

let upserted = MyCollection::entry("unique name", &db)
    .update_with(|existing: &mut MyCollection| {
        existing.rank += 1;
    })
    .or_insert_with(MyCollection::default)
    .execute()?
    .unwrap();
println!("Rank: {:?}", upserted.contents.rank);

Associated Types

type ByNameView: SerializedView<Key = String>

The name view defined for the collection.

Provided methods

fn load<'name, N: Nameable<'name, Self::PrimaryKey> + Send + Sync, C: Connection>(
id: N,
connection: &C
) -> Result<Option<CollectionDocument<Self>>, Error> where
Self: SerializedCollection + Sized + 'static,

Gets a CollectionDocument with id from connection.

fn load_async<'name, 'life0, 'async_trait, N, C>(
id: N,
connection: &'life0 C
) -> Pin<Box<dyn Future<Output = Result<Option<CollectionDocument<Self>>, Error>> + Send + 'async_trait>> where
Self: SerializedCollection + Sized + 'static,
'name: 'async_trait,
N: 'async_trait + Nameable<'name, Self::PrimaryKey> + Send + Sync,
C: 'async_trait + AsyncConnection,
'life0: 'async_trait,
Self: Send + 'async_trait,

Gets a CollectionDocument with id from connection.

fn entry<'connection, 'name, N: Into<NamedReference<'name, Self::PrimaryKey>> + Send + Sync, C: Connection>(
id: N,
connection: &'connection C
) -> Entry<'connection, 'name, C, Self, (), ()> where
Self: SerializedCollection + Sized,

Gets a CollectionDocument with id from connection.

fn entry_async<'connection, 'name, N: Into<NamedReference<'name, Self::PrimaryKey>> + Send + Sync, C: AsyncConnection>(
id: N,
connection: &'connection C
) -> AsyncEntry<'connection, 'name, C, Self, (), ()> where
Self: SerializedCollection + Sized,

Gets a CollectionDocument with id from connection.

fn load_document<'name, N: Nameable<'name, Self::PrimaryKey> + Send + Sync, C: Connection>(
name: N,
connection: &C
) -> Result<Option<OwnedDocument>, Error> where
Self: SerializedCollection + Sized,

Loads a document from this collection by name, if applicable. Return Ok(None) if unsupported.

fn load_document_async<'name, 'life0, 'async_trait, N, C>(
name: N,
connection: &'life0 C
) -> Pin<Box<dyn Future<Output = Result<Option<OwnedDocument>, Error>> + Send + 'async_trait>> where
Self: SerializedCollection + Sized,
'name: 'async_trait,
N: 'async_trait + Nameable<'name, Self::PrimaryKey> + Send + Sync,
C: 'async_trait + AsyncConnection,
'life0: 'async_trait,
Self: Send + 'async_trait,

Loads a document from this collection by name, if applicable. Return Ok(None) if unsupported.