Trait bonsaidb_core::schema::SerializedCollection

pub trait SerializedCollection: Collection {
    type Contents: Send + Sync;
    type Format: Format<'static, Self::Contents> + OwnedDeserializer<Self::Contents>;
    fn format() -> Self::Format;

    fn deserialize(data: &[u8]) -> Result<Self::Contents, Error> { ... }

    fn serialize(item: &Self::Contents) -> Result<Vec<u8>, Error> { ... }

    fn get<'life0, 'async_trait, C: Connection>(
        id: u64, 
        connection: &'life0 C
    ) -> Pin<Box<dyn Future<Output = Result<Option<CollectionDocument<Self>>, Error>> + Send + 'async_trait>>
    where
        Self: Sized,
        C: 'async_trait,
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }

    fn push<'life0, 'async_trait, Cn: Connection>(
        contents: Self::Contents, 
        connection: &'life0 Cn
    ) -> Pin<Box<dyn Future<Output = Result<CollectionDocument<Self>, InsertError<Self::Contents>>> + Send + 'async_trait>>
    where
        Self: Sized + 'static,
        Self::Contents: 'async_trait,
        Cn: 'async_trait,
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }

    fn push_into<'life0, 'async_trait, Cn: Connection>(
        self, 
        connection: &'life0 Cn
    ) -> Pin<Box<dyn Future<Output = Result<CollectionDocument<Self>, InsertError<Self>>> + Send + 'async_trait>>
    where
        Self: SerializedCollection<Contents = Self> + Sized + 'static,
        Cn: 'async_trait,
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
}

A collection that knows how to serialize and deserialize documents to an associated type.

Associated Types

type Contents: Send + Sync

The type of the contents stored in documents in this collection.

type Format: Format<'static, Self::Contents> + OwnedDeserializer<Self::Contents>

The serialization format for this collection.

Required methods

fn format() -> Self::Format

Returns the configured instance of Self::Format.

Provided methods

fn deserialize(data: &[u8]) -> Result<Self::Contents, Error>

Deserialize data as Self::Contents using this collection’s format.

fn serialize(item: &Self::Contents) -> Result<Vec<u8>, Error>

Serialize item using this collection’s format.

fn get<'life0, 'async_trait, C: Connection>(
id: u64,
connection: &'life0 C
) -> Pin<Box<dyn Future<Output = Result<Option<CollectionDocument<Self>>, Error>> + Send + 'async_trait>> where
Self: Sized,
C: 'async_trait,
'life0: 'async_trait,
Self: Send + 'async_trait,

Gets a CollectionDocument with id from connection.

fn push<'life0, 'async_trait, Cn: Connection>(
contents: Self::Contents,
connection: &'life0 Cn
) -> Pin<Box<dyn Future<Output = Result<CollectionDocument<Self>, InsertError<Self::Contents>>> + Send + 'async_trait>> where
Self: Sized + 'static,
Self::Contents: 'async_trait,
Cn: 'async_trait,
'life0: 'async_trait,
Self: Send + 'async_trait,

Pushes this value into the collection, returning the created document.

fn push_into<'life0, 'async_trait, Cn: Connection>(
self,
connection: &'life0 Cn
) -> Pin<Box<dyn Future<Output = Result<CollectionDocument<Self>, InsertError<Self>>> + Send + 'async_trait>> where
Self: SerializedCollection<Contents = Self> + Sized + 'static,
Cn: 'async_trait,
'life0: 'async_trait,
Self: Send + 'async_trait,

Pushes this value into the collection, returning the created document.