Trait bonsaidb_core::schema::view::Serialized

pub trait Serialized: Send + Sync + Debug {
    fn collection(&self) -> CollectionName;
    fn unique(&self) -> bool;
    fn lazy(&self) -> bool;
    fn version(&self) -> u64;
    fn view_name(&self) -> ViewName;
    fn map(
        &self,
        document: &BorrowedDocument<'_>
    ) -> Result<Vec<Serialized>, Error>;
    fn reduce(
        &self,
        mappings: &[(&[u8], &[u8])],
        rereduce: bool
    ) -> Result<Vec<u8>, Error>;

    fn eager(&self) -> bool { ... }
}

Wraps a View with serialization to erase the associated types

Required Methods

fn collection(&self) -> CollectionName

Wraps returing <View::Collection as Collection>::collection_name()

fn unique(&self) -> bool

Wraps ViewSchema::unique

fn lazy(&self) -> bool

Wraps ViewSchema::lazy

fn version(&self) -> u64

Wraps ViewSchema::version

fn view_name(&self) -> ViewName

Wraps View::view_name

fn map(&self, document: &BorrowedDocument<'_>) -> Result<Vec<Serialized>, Error>

Wraps ViewSchema::map

fn reduce(
&self,
mappings: &[(&[u8], &[u8])],
rereduce: bool
) -> Result<Vec<u8>, Error>

Wraps ViewSchema::reduce

Provided Methods

fn eager(&self) -> bool

Returns true if this view should be eagerly updated during document updates.