Trait bonsaidb_core::schema::view::SerializedView

pub trait SerializedView: View {
    type Format: OwnedDeserializer<Self::Value>;

    fn format() -> Self::Format;

    fn deserialize(data: &[u8]) -> Result<Self::Value, Error> { ... }
    fn serialize(item: &Self::Value) -> Result<Vec<u8>, Error> { ... }
    fn entries<Database: Connection>(
        database: &Database
    ) -> View<'_, Database, Self, Self::Key> { ... }
    fn entries_async<Database: AsyncConnection>(
        database: &Database
    ) -> AsyncView<'_, Database, Self, Self::Key> { ... }
}

A View with additional tyes and logic to handle serializing view values.

Required Associated Types

type Format: OwnedDeserializer<Self::Value>

The serialization format for this view.

Required Methods

fn format() -> Self::Format

Returns the configured instance of Self::Format.

Provided Methods

fn deserialize(data: &[u8]) -> Result<Self::Value, Error>

Deserialize data as Self::Value using this views’s format.

fn serialize(item: &Self::Value) -> Result<Vec<u8>, Error>

Serialize item using this views’s format.

fn entries<Database: Connection>(
database: &Database
) -> View<'_, Database, Self, Self::Key>

Returns a builder for a view query or view reduce.

fn entries_async<Database: AsyncConnection>(
database: &Database
) -> AsyncView<'_, Database, Self, Self::Key>

Returns a builder for a view query or view reduce.