Trait bonsaidb_core::schema::view::View

pub trait View: Send + Sync + Debug + 'static {
    type Collection: Collection;
    type Key: Key + 'static;
    type Value: Send + Sync;
    fn name(&self) -> Result<Name, InvalidNameError>;

    fn view_name(&self) -> Result<ViewName, InvalidNameError> { ... }
}

A mechanism for accessing mapped and/or reduced data from a Collection.

See the user guide for a walkthrough of how views work.

Associated Types

type Collection: Collection

The collection this view belongs to

type Key: Key + 'static

The key for this view.

type Value: Send + Sync

An associated type that can be stored with each entry in the view.

Required methods

fn name(&self) -> Result<Name, InvalidNameError>

The name of the view. Must be unique per collection.

Provided methods

fn view_name(&self) -> Result<ViewName, InvalidNameError>

The namespaced name of the view.