Trait bonsaidb_core::schema::view::ViewSchema

pub trait ViewSchema: Send + Sync + Debug + 'static {
    type View: SerializedView;
    fn map(&self, document: &Document) -> ViewMapResult<Self::View>;

    fn unique(&self) -> bool { ... }

    fn version(&self) -> u64 { ... }

    fn reduce(
        &self, 
        mappings: &[ViewMappedValue<Self::View>], 
        rereduce: bool
    ) -> Result<<Self::View as View>::Value, Error> { ... }
}

The implementation of Map/Reduce for a View.

See the user guide for a walkthrough of how views work.

Associated Types

type View: SerializedView

The view this schema is defined for.

Required methods

fn map(&self, document: &Document) -> ViewMapResult<Self::View>

The map function for this view. This function is responsible for emitting entries for any documents that should be contained in this View. If None is returned, the View will not include the document. See the user guide’s chapter on views for more information on how map works.

Provided methods

fn unique(&self) -> bool

If true, no two documents may emit the same key. Unique views are updated when the document is saved, allowing for this check to be done atomically. When a document is updated, all unique views will be updated, and if any of them fail, the document will not be allowed to update and an Error::UniqueKeyViolation will be returned.

fn version(&self) -> u64

The version of the view. Changing this value will cause indexes to be rebuilt.

fn reduce(
&self,
mappings: &[ViewMappedValue<Self::View>],
rereduce: bool
) -> Result<<Self::View as View>::Value, Error>

Returns a value that is produced by reducing a list of mappings into a single value. If rereduce is true, the values contained in the mappings have already been reduced at least one time. If Err(Error::ReduceUnimplemented) is returned, queries that ask for a reduce operation will return an error. See the user guide’s chapter on views for more information on how reduce works.