Trait bonsaidb_core::key::IntoPrefixRange

pub trait IntoPrefixRange: Sized {
    fn into_prefix_range(self) -> Range<Self>;
}

A type that can be used as a prefix range in range-based queries.

Required methods

fn into_prefix_range(self) -> Range<Self>

Returns the value as a prefix-range, which will match all values that start with self.