Module bonsaidb_core::transaction

Types for executing transactions.

Structs

ChangedDocument

A record of a changed document.

ChangedKey

A record of a changed KeyValue entry.

DocumentChanges

A list of changed documents.

DocumentChangesIntoIter

A draining iterator over ChangedDocuments.

DocumentChangesIter

An iterator over DocumentChanges.

Executed

Details about an executed transaction.

Operation

A single operation performed on a Collection.

Transaction

A list of operations to execute as a single unit. If any operation fails, all changes are aborted. Transactions are ACID-compliant. ACID stands for:

Enums

Changes

A list of changes.

Command

A command to execute within a Collection.

OperationResult

Information about the result of each Operation in a transaction.